+++ plone/app/event/base.py
@@ -12,7 +12,6 @@ from datetime import timedelta
 from persistent.dict import PersistentDict
 from plone.app.event.interfaces import IEventSettings
 from plone.app.event.interfaces import ISO_DATE_FORMAT
-from plone.app.event.vocabularies import replacement_zones
 from plone.app.layout.navigation.root import getNavigationRootObject
 from plone.event.interfaces import IEvent
 from plone.event.interfaces import IEventAccessor
@@ -38,6 +37,25 @@ import pytz
 
 DEFAULT_END_DELTA = 1  # hours
 FALLBACK_TIMEZONE = 'UTC'
+
+# Sync strategies
+SYNC_NONE = 0
+SYNC_KEEP_NEWER = 1
+SYNC_KEEP_MINE = 2
+SYNC_KEEP_THEIRS = 3
+
+# Map for ambiguous timezone abbreviations to their most common non-ambigious
+# timezone name. E.g CST is ambiguous and is used for U.S./Canada Central
+# Standard Time, Australian Central Standard Time, China Standard Time.
+# TODO: incomplete map.
+# TODO: do we need this at all or shouldn't we just fail with ambiguous
+#       timezones?
+replacement_zones = {
+    'CET': 'Europe/Vienna',    # Central European Time
+    'MET': 'Europe/Vienna',    # Middle European Time
+    'EET': 'Europe/Helsinki',  # East European Time
+    'WET': 'Europe/Lisbon',    # West European Time
+}
 
 
 # RETRIEVE EVENTS